+++ indexcov/functional-tests.sh
@@ -36,3 +36,6 @@ run check_single_sample ./goleft_test indexcov -d /tmp/tt sample_name_0001.bam
 assert_exit_code 0
 assert_in_stderr "not plotting"
 
+run check_sex ./goleft_test indexcov -d /tmp/tt --sex "X,Y" sample_name_0001.bam
+assert_exit_code 0
+